refactor(room-escape): 리팩토링
